Add test2.html
Adding test2.html to js folder to serve as code comparison to try to figure out why App.js initMap is not rendering the map. App.js code was written without order and used actual addresses for geocoder, and test2.js is coded in order and the addresses is LatLng, and the map renders.